Rename middlewares -> middleware, add BouncerMiddleware.
Before each plugin middleware had to hit the database to get its
configuration. Now, BouncerMiddleware checks the request HTTP_HOST
header and get all plugins that apply to that host and attach all
related plugins to request.META, this way it only has to hit the
database once per request to get plugin configuration.

All other plugin middleware now have to get their config from
request.META['BOUNCER_PLUGINS']
