+++ api_bouncer/middleware/ip_restriction.py
@@ -2,8 +2,6 @@ import ipaddress
 
 from django.http import JsonResponse
 
-from ..models import Plugin
-
 
 def get_client_ip(request):
     x_forwarded_for = request.META.get('HTTP_X_FORWARDED_FOR')
@@ -19,21 +17,15 @@ class IpRestrictionMiddleware(object):
         self.get_response = get_response
 
     def __call__(self, request):
-        host = request.META.get('HTTP_HOST')
         consumer_id = request.META.get('HTTP_CONSUMER_ID')
-
-        plugin_conf = Plugin.objects.filter(
-            api__hosts__contains=[host],
-            name='ip-restriction'
-        ).first()
+        config = request.META['BOUNCER_PLUGINS'].get('ip-restriction')
 
         if (
-            plugin_conf and (
-                not plugin_conf.config.get('consumer_id') or
-                plugin_conf.config.get('consumer_id') == consumer_id
+            config and (
+                not config.get('consumer_id') or
+                config.get('consumer_id') == consumer_id
             )
         ):
-            config = plugin_conf.config
             whitelist = config['whitelist']
             blacklist = config['blacklist']
             client_ip = get_client_ip(request)
